+++ packages/agent/CHANGELOG.md
@@ -1,5 +1,11 @@
 # mycoder-agent
 
+## 0.2.0
+
+### Minor Changes
+
+- Add token caching, better user input handling, token usage logging (--tokenUsage), the ability to see the browser (--headless=false), and log prefixes with emojis.
+
 ## 0.1.3
 
 ### Patch Changes